+++ docs/api/data.batchify.rst
@@ -27,6 +27,13 @@ Language Modeling
     CorpusBPTTBatchify
     StreamBPTTBatchify
 
+Embedding Training
+------------------
+
+.. autosummary::
+    :nosignatures:
+
+    EmbeddingCenterContextBatchify
 
 API Reference
 -------------
@@ -34,4 +41,4 @@ API Reference
 .. automodule:: gluonnlp.data.batchify
    :members:
    :imported-members:
-   :special-members: __call__
+   :special-members: __call__, __iter__
